1. Добавлен модуль :features и в нем :settings и :category
