+++ .idea/deploymentTargetSelector.xml
@@ -11,6 +11,9 @@
       <SelectionState runConfigName="navigation">
         <option name="selectionMode" value="DROPDOWN" />
       </SelectionState>
+      <SelectionState runConfigName="features">
+        <option name="selectionMode" value="DROPDOWN" />
+      </SelectionState>
     </selectionStates>
   </component>
 </project>
