+++ app/src/main/java/com/example/yandex_school_app/di/AppComponent.kt
@@ -1,5 +1,6 @@
 package com.example.yandex_school_app.di
 
+import com.example.common.di.ApplicationScope
 import com.example.yandex_school_app.MainActivity
 import com.example.yandex_school_app.di.module.DataModule
 import com.example.yandex_school_app.di.module.DomainModule
