--- app/src/main/java/com/example/yandex_school_app/di/ApplicationScope.kt
@@ -1,7 +1,0 @@
-package com.example.yandex_school_app.di
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-annotation class ApplicationScope
