+++ app/src/main/java/com/example/yandex_school_app/di/module/DomainModule.kt
@@ -2,11 +2,11 @@ package com.example.yandex_school_app.di.module
 
 import com.example.common.data.repository.TransactionRepositoryImpl
 import com.example.common.domain.repository.TransactionRepository
-import com.example.yandex_school_app.di.ApplicationScope
+import com.example.common.di.ApplicationScope
 import com.example.yandex_school_app.features.cash_account.data.repository.AccountRepositoryImpl
 import com.example.yandex_school_app.features.cash_account.domain.repository.AccountRepository
-import com.example.yandex_school_app.features.category.data.repository.CategoryRepositoryImpl
-import com.example.yandex_school_app.features.category.domain.repository.CategoryRepository
+import com.example.category.data.repository.CategoryRepositoryImpl
+import com.example.category.domain.repository.CategoryRepository
 import dagger.Binds
 import dagger.Module
