+++ app/src/main/java/com/example/yandex_school_app/features/cash_account/data/datasource/remote/AccountRemoteDataSource.kt
@@ -1,7 +1,7 @@
 package com.example.yandex_school_app.features.cash_account.data.datasource.remote
 
 import com.example.common.data.network.ResponseTemplate
-import com.example.yandex_school_app.features.cash_account.domain.entity.AccountDomain
+import com.example.common.domain.entity.AccountDomain
 
 interface AccountRemoteDataSource {
     suspend fun getAllCashAccount(): ResponseTemplate<List<AccountDomain>>
