--- app/src/main/java/com/example/yandex_school_app/features/cash_account/domain/entity/AccountDomain.kt
@@ -1,8 +1,0 @@
-package com.example.yandex_school_app.features.cash_account.domain.entity
-
-data class AccountDomain(
-    val id: Int,
-    val name: String,
-    val balance: String,
-    val currency: String
-)
