+++ app/src/main/java/com/example/yandex_school_app/features/cash_account/presentation/ui/CreateCashAccount.kt
@@ -24,7 +24,7 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import com.example.yandex_school_app.MainActivity
 import com.example.common.presentation.ToastController
-import com.example.yandex_school_app.features.cash_account.domain.entity.AccountDomain
+import com.example.common.domain.entity.AccountDomain
 import com.example.yandex_school_app.features.cash_account.domain.entity.Currency
 import com.example.yandex_school_app.features.cash_account.presentation.AccountViewModel
