--- app/src/main/java/com/example/yandex_school_app/features/category/data/network/CategoryApiClient.kt
@@ -1,15 +1,0 @@
-package com.example.yandex_school_app.features.category.data.network
-
-import com.example.common.data.network.BaseUrl
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object CategoryApiClient {
-    val categoryApiService: CategoryApiService by lazy {
-        Retrofit.Builder()
-            .baseUrl(BaseUrl.getUrl())
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(CategoryApiService::class.java)
-    }
-}
