--- app/src/main/java/com/example/yandex_school_app/features/category/data/repository/CategoryRepositoryImpl.kt
@@ -1,15 +1,0 @@
-package com.example.yandex_school_app.features.category.data.repository
-
-import com.example.common.data.network.ResponseTemplate
-import com.example.yandex_school_app.features.category.data.datasource.remote.CategoryRemoteDataSource
-import com.example.common.domain.entity.CategoryDomain
-import com.example.yandex_school_app.features.category.domain.repository.CategoryRepository
-import javax.inject.Inject
-
-class CategoryRepositoryImpl @Inject constructor(
-    private val categoryRemoteDataSource: CategoryRemoteDataSource,
-) : CategoryRepository {
-    override suspend fun getCategories(): ResponseTemplate<List<CategoryDomain>> {
-        return categoryRemoteDataSource.getCategories()
-    }
-}
