--- app/src/main/java/com/example/yandex_school_app/features/category/domain/repository/CategoryRepository.kt
@@ -1,8 +1,0 @@
-package com.example.yandex_school_app.features.category.domain.repository
-
-import com.example.common.data.network.ResponseTemplate
-import com.example.common.domain.entity.CategoryDomain
-
-interface CategoryRepository {
-    suspend fun getCategories(): ResponseTemplate<List<CategoryDomain>>
-}
