--- app/src/main/java/com/example/yandex_school_app/features/category/domain/usecase/CategoryUseCase.kt
@@ -1,10 +1,0 @@
-package com.example.yandex_school_app.features.category.domain.usecase
-
-import com.example.yandex_school_app.features.category.domain.repository.CategoryRepository
-import javax.inject.Inject
-
-class CategoryUseCase @Inject constructor(
-    private val categoriesRepository: CategoryRepository
-) {
-    suspend fun getCategories() = categoriesRepository.getCategories()
-}
