+++ app/src/main/java/com/example/yandex_school_app/features/income/presentation/viewmodel/IncomeViewModel.kt
@@ -6,7 +6,7 @@ import com.example.common.data.network.ResponseTemplate
 import com.example.common.domain.entity.TransactionDomain
 import com.example.common.domain.usecase.TransactionUseCase
 import com.example.common.presentation.ToastController
-import com.example.yandex_school_app.di.AccountManager
+import com.example.common.manager.AccountManager
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
