+++ common/src/main/java/com/example/common/di/ApplicationScope.kt
@@ -1,0 +1,7 @@
+package com.example.common.di
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class ApplicationScope
