+++ common/src/main/java/com/example/common/domain/entity/AccountDomain.kt
@@ -1,0 +1,8 @@
+package com.example.common.domain.entity
+
+data class AccountDomain(
+    val id: Int,
+    val name: String,
+    val balance: String,
+    val currency: String
+)
