+++ features/category/src/main/java/com/example/category/data/network/CategoryApiService.kt
@@ -1,0 +1,14 @@
+package com.example.category.data.network
+
+import com.example.common.data.network.model.CategoryNetwork
+import com.example.common.data.network.BaseUrl
+import retrofit2.Call
+import retrofit2.http.GET
+import retrofit2.http.Header
+
+interface CategoryApiService {
+    @GET("categories")
+    fun getCategories(
+        @Header("Authorization") token: String = BaseUrl.getToken()
+    ): Call<List<CategoryNetwork>>
+}
