+++ features/category/src/main/java/com/example/category/domain/repository/CategoryRepository.kt
@@ -1,0 +1,8 @@
+package com.example.category.domain.repository
+
+import com.example.common.data.network.ResponseTemplate
+import com.example.common.domain.entity.CategoryDomain
+
+interface CategoryRepository {
+    suspend fun getCategories(): ResponseTemplate<List<CategoryDomain>>
+}
