+++ settings.gradle.kts
@@ -22,4 +22,7 @@ dependencyResolutionManagement {
 rootProject.name = "Yandex_School_App"
 include(":app")
 include(":navigation")
+include(":features")
 include(":common")
+include(":features:settings")
+include(":features:category")
